+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Unreleased
-
-# v1.1.0
 ## 🚀 Features
+- `create` has option `--run-container` to execute the command for CommandLineTool creation with provided container
+- `execute local --apptainer` or `execute local --singularity` to use apptainer/singularity for local execution
+  
+# v1.1.0
 - `create` now has `--env` option to load in an .env file
 - made inputs and outputs prefix optional in `connect`